requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
easyocr==1.7.1
fastapi==0.104.1
filelock==3.13.1
fsspec==2023.10.0
h11==0.14.0
idna==3.4
imageio==2.31.6
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.1
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.1
pyclipper==1.3.0.post5
pydantic==2.4.2
pydantic_core==2.10.1
pytesseract==0.3.10
python-bidi==0.4.2
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tesseract==0.1.3
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
